+++ app/src/main/java/woowacourse/shopping/shoppingcart/CartPresenter.kt
@@ -5,6 +5,8 @@ import androidx.lifecycle.MutableLiveData
 import model.Cart
 import woowacourse.shopping.database.ShoppingCache
 import woowacourse.shopping.model.CartProductUiModel
+import woowacourse.shopping.shoppingcart.pagination.CartPage
+import woowacourse.shopping.shoppingcart.pagination.CartPageHandler
 import woowacourse.shopping.util.toCartProductUiModel
 
 class CartPresenter(
