+++ app/src/main/java/woowacourse/shopping/shoppingcart/adapter/CartProductCountPickerListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.shoppingcart
+package woowacourse.shopping.shoppingcart.adapter
 
 interface CartProductCountPickerListener {
